Refactor path to be overridden
Calling scripts can set `bgit.REPO_PATH` so it doesn't need to be
specified as an argument each time. If the value isn't overridden, a git
repo will be created on the Desktop in an `errors` folder so I'll know
something is broken and which script it's broken in based on the commit
message.